requirements.txt
absl-py==0.12.0
aiosmtplib==1.1.6
alembic==1.6.5
amqp==5.0.9
anyio==3.4.0
asgiref==3.4.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.3.0
billiard==3.6.4.0
blinker==1.4
cached-property==1.5.2
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.10
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
confluent-kafka==1.7.0
dill==0.3.4
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.0
fastapi-mail==1.0.2
flower==1.0.0
future==0.18.2
googleapis-common-protos==1.54.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.13.7
httptools==0.3.0
httpx==0.19.0
humanize==3.13.1
idna==3.3
importlib-metadata==2.1.2
importlib-resources==5.4.0
Jinja2==3.0.3
kafka==1.3.5
kafka-python==2.0.2
kazoo==2.5.0
kombu==5.2.2
loguru==0.5.3
Mako==1.1.6
MarkupSafe==2.0.1
msgpack==1.0.3
numpy==1.21.5
packaging==21.3
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.24
protobuf==3.19.1
psycopg2-binary==2.9.1
pydantic==1.8.2
pykafka==2.8.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
python-editor==1.0.4
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
requests==2.27.1
rfc3986==1.5.0
send-email==201212
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.20
starlette==0.16.0
tabulate==0.8.9
tensorflow-datasets==4.4.0
tensorflow-metadata==1.5.0
termcolor==1.1.0
tornado==6.1
tqdm==4.62.3
typeguard==2.12.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.16.0
uvloop==0.16.0
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.1
zipp==3.7.0
zookeeper==1.3.2
